requirements.txt
mcp==1.6.0
openai==1.76.0
openai-agents==0.0.13
python-dotenv==1.1.0
requests==2.32.3
aiohttp==3.11.18
asyncio==3.4.3
colorama==0.4.6
python-dateutil==2.9.0.post0
sentence-transformers==4.1.0
faiss-cpu==1.11.0
huggingface_hub==0.30.2
tqdm==4.67.1
scikit-learn==1.6.1
scipy==1.15.2
numpy==2.2.5
transformers==4.51.3
pydantic==2.11.3
pydantic-settings==2.9.1
griffe==1.7.3
anyio==4.9.0
httpx==0.28.1
httpx-sse==0.4.0
sse-starlette==2.3.3
starlette==0.46.2
certifi==2025.4.26
charset-normalizer==3.4.1
idna==3.10
urllib3==2.4.0
typing-extensions==4.13.2
markdown==3.8
fastapi
uvicorn
playwright==1.42.0
greenlet==3.0.3
pyee==11.0.1
openai>=1.12.0
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.14
